+++ src/test/java/com/getindata/connectors/http/internal/sink/httpclient/BatchRequestSubmitterTest.java
@@ -1,6 +1,7 @@
 package com.getindata.connectors.http.internal.sink.httpclient;
 
 import java.net.http.HttpClient;
+import java.util.Collections;
 import java.util.List;
 import java.util.Properties;
 import java.util.concurrent.CompletableFuture;
@@ -8,6 +9,7 @@ import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 import java.util.stream.Stream;
 
+import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
@@ -15,6 +17,7 @@ import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.MethodSource;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
@@ -92,5 +95,28 @@ class BatchRequestSubmitterTest {
         );
 
         verify(mockHttpClient, times(expectedNumberOfBatchRequests)).sendAsync(any(), any());
+    }
+
+    @Test
+    public void shouldHandleEmptyBatch() {
+
+        Properties properties = new Properties();
+        properties.setProperty(
+                HttpConnectorConfigConstants.SINK_HTTP_BATCH_REQUEST_SIZE,
+                String.valueOf(50)
+        );
+
+        BatchRequestSubmitter submitter = new BatchRequestSubmitter(
+                properties,
+                new String[0],
+                mockHttpClient
+        );
+
+        List<CompletableFuture<JavaNetHttpResponseWrapper>> futures = submitter.submit(
+                "http://hello.pl",
+                Collections.emptyList()
+        );
+
+        assertEquals(futures, Collections.emptyList());
     }
 }